Move ZeroGuid definition
