+++ packages/sdk/src/constants.ts
@@ -5,6 +5,8 @@
 
 // tslint:disable:variable-name
 
+export const ZeroGuid = '00000000-0000-0000-0000-000000000000';
+
 /** @hidden */
 export const HTTPHeaders = {
     SessionID: 'x-ms-mixed-reality-extension-sessionid',
