+++ packages/sdk/src/types/runtime/assets/material.ts
@@ -4,8 +4,8 @@
  */
 
 import { Asset, AssetLike, AssetManager, Texture } from '.';
+import { ZeroGuid } from '../../../constants';
 import { Color3, Color4, Vector2, Vector2Like } from '../../../math';
-import { ZeroGuid } from '../../../utils/deterministicGuids';
 import observe from '../../../utils/observe';
 import readPath from '../../../utils/readPath';
 import { InternalAsset } from '../../internal/asset';
